Django/API/products/products/serializers/serializers_attributename.py
from rest_framework import serializers
from ...models import AttributeName

class AttributeNameSerializer(serializers.ModelSerializer):
    class Meta:
        model = AttributeName
        fields = ['id', 'name']
